Move new esbuild plugin to private package
