+++ packages/esbuild-plugin-next/package.json
@@ -1,0 +1,24 @@
+{
+  "private": true,
+  "name": "@vanilla-extract/esbuild-plugin-next",
+  "version": "0.0.0",
+  "description": "Zero-runtime Stylesheets-in-TypeScript",
+  "main": "dist/vanilla-extract-esbuild-plugin-next.cjs.js",
+  "module": "dist/vanilla-extract-esbuild-plugin-next.esm.js",
+  "files": [
+    "/dist"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/vanilla-extract-css/vanilla-extract.git",
+    "directory": "packages/esbuild-plugin-next"
+  },
+  "author": "SEEK",
+  "license": "MIT",
+  "dependencies": {
+    "@vanilla-extract/integration": "^6.0.2"
+  },
+  "devDependencies": {
+    "esbuild": "0.17.6"
+  }
+}
