+++ packages/integration/src/index.ts
@@ -6,7 +6,7 @@ export {
 export { getSourceFromVirtualCssFile } from './virtualFile';
 export { getPackageInfo } from './packageInfo';
 export { compile, vanillaExtractTransformPlugin } from './compile';
-export { createCompiler as unstable_createCompiler } from './compiler';
+export { createCompiler } from './compiler';
 export { hash } from './hash';
 export { addFileScope } from './addFileScope';
 export { serializeCss, deserializeCss } from './serialize';
@@ -17,4 +17,4 @@ export * from './filters';
 export type { IdentifierOption } from './types';
 export type { PackageInfo } from './packageInfo';
 export type { CompileOptions } from './compile';
-export type { CreateCompilerOptions as Unstable_CreateCompilerOptions } from './compiler';
+export type { CreateCompilerOptions } from './compiler';
