+++ test-helpers/package.json
@@ -18,6 +18,7 @@
     "@types/mini-css-extract-plugin": "^1.2.2",
     "@types/webpack-dev-server": "^3.11.1",
     "@vanilla-extract/esbuild-plugin": "*",
+    "@vanilla-extract/esbuild-plugin-next": "*",
     "@vanilla-extract/parcel-transformer": "*",
     "@vanilla-extract/vite-plugin": "*",
     "@vanilla-extract/webpack-plugin": "*",
